api/requirements.txt
flask==1.0.*
aws-wsgi==0.0.*
simplejson==3.*
google-api-python-client==1.7.*